+++ test/clutch-bench.el
@@ -4,7 +4,6 @@
 
 ;; Author: Lucius Chen <chenyh572@gmail.com>
 ;; Maintainer: Lucius Chen <chenyh572@gmail.com>
-;; Version: 0.1.0
 ;; URL: https://github.com/LuciusChen/clutch
 
 ;; This file is part of clutch.